# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MaterialPackageCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "copy_generation_run_id": {
            "format": "uuid",
            "title": "Copy Generation Run Id",
            "type": "string"
          },
          "reviewer": {
            "default": "internal",
            "maxLength": 120,
            "minLength": 1,
            "title": "Reviewer",
            "type": "string"
          }
        },
        "required": [
          "copy_generation_run_id"
        ],
        "title": "MaterialPackageCreateRequest",
        "type": "object"
      },
...
      "MaterialPackageListResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/MaterialPackageSummaryResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "MaterialPackageListResponse",
        "type": "object"
      },
      "MaterialPackageResponse": {
        "properties": {
          "audit": {
            "additionalProperties": true,
            "title": "Audit",
            "type": "object"
          },
          "brand_bindings": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Brand Bindings",
            "type": "array"
          },
          "business_date": {
            "title": "Business Date",
            "type": "string"
          },
          "content_slot": {
            "anyOf": [
              {
                "enum": [
                  "morning",
                  "noon",
                  "evening"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Slot"
          },
          "copy": {
            "additionalProperties": true,
            "title": "Copy",
            "type": "object"
          },
          "copy_generation_run_id": {
            "format": "uuid",
            "title": "Copy Generation Run Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "detail_url": {
            "title": "Detail Url",
            "type": "string"
          },
          "download_url": {
            "title": "Download Url",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image": {
            "$ref": "#/components/schemas/ImageArtifactResponse"
          },
          "ordinal": {
            "anyOf": [
              {
                "maximum": 3.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ordinal"
          },
          "package_version": {
            "title": "Package Version",
            "type": "integer"
          },
          "review_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Note"
          },
          "review_status": {
            "enum": [
              "pending",
              "approved",
              "rejected"
            ],
            "title": "Review Status",
            "type": "string"
          },
          "review_url": {
            "title": "Review Url",
            "type": "string"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "sources": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "enum": [
              "queued",
              "ready",
              "awaiting_manual_use",
              "completed",
              "rejected",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "target_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target At"
          },
          "topic": {
            "additionalProperties": true,
            "title": "Topic",
            "type": "object"
          },
          "validation": {
            "additionalProperties": true,
            "title": "Validation",
            "type": "object"
          },
          "versions": {
            "additionalProperties": true,
            "title": "Versions",
            "type": "object"
          }
        },
        "required": [
          "id",
          "copy_generation_run_id",
          "status",
          "review_status",
          "business_date",
          "created_at",
          "detail_url",
          "package_version",
          "topic",
          "copy",
          "sources",
          "brand_bindings",
          "validation",
          "audit",
          "versions",
          "image",
          "review_note",
          "reviewed_at",
          "review_url",
          "download_url"
        ],
        "title": "MaterialPackageResponse",
        "type": "object"
      },
...
      "MaterialReviewRequest": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "enum": [
              "approved",
              "rejected"
            ],
            "title": "Decision",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "reviewer": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Reviewer",
            "type": "string"
          }
        },
        "required": [
          "decision",
          "reviewer"
        ],
        "title": "MaterialReviewRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Internal API for authoritative-source acquisition, factual governance, and auditable event organization, deterministic daily topic selection, and private brand knowledge ingestion with internal copy-generation context retrieval. It does not expose public brand search, publishing, or arbitrary URL-fetch capabilities.",
    "title": "Edu AI Lead Agent API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/material-packages": {
      "get": {
        "operationId": "list_material_packages_api_v1_material_packages_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialPackageListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Material Packages",
        "tags": [
          "material-packages"
        ]
      },
      "post": {
        "operationId": "generate_material_package_api_v1_material_packages_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialPackageCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialPackageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Material Package",
        "tags": [
          "material-packages"
        ]
      }
    },
    "/api/v1/material-packages/{package_id}": {
      "get": {
        "operationId": "read_material_package_api_v1_material_packages__package_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "package_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Package Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialPackageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Material Package",
        "tags": [
          "material-packages"
        ]
      }
    },
    "/api/v1/material-packages/{package_id}/download": {
      "get": {
        "description": "Return a safe, attachment-friendly JSON package without object-store internals.",
        "operationId": "download_material_package_api_v1_material_packages__package_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "package_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Package Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialPackageDownloadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Material Package",
        "tags": [
          "material-packages"
        ]
      }
    },
    "/api/v1/material-packages/{package_id}/image": {
      "get": {
        "operationId": "download_material_package_image_api_v1_material_packages__package_id__image_get",
        "parameters": [
          {
            "in": "path",
            "name": "package_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Package Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Material Package Image",
        "tags": [
          "material-packages"
        ]
      }
    },
    "/api/v1/material-packages/{package_id}/image/retry": {
      "post": {
        "operationId": "retry_material_package_image_route_api_v1_material_packages__package_id__image_retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "package_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Package Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialPackageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Material Package Image Route",
        "tags": [
          "material-packages"
        ]
      }
    },
    "/api/v1/material-packages/{package_id}/review": {
      "post": {
        "operationId": "review_material_package_route_api_v1_material_packages__package_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "package_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Package Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialPackageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Material Package Route",
        "tags": [
          "material-packages"
        ]
      }
    },
...
  }
}